+++ src/emu/arithmetic/inr.rs
@@ -9,7 +9,7 @@ pub fn execute_inr(state: &mut State, inr: INR) {
     match inr {
         INR::IncrementRegister { r } => {
             let lhs = state.get_register(&r) as i8;
-            let rhs = 1 as i8;
+            let rhs = 1_i8;
 
             let (result, carried) = lhs.overflowing_add(rhs);
             let flags = get_flags(lhs, rhs, result, carried);
@@ -26,7 +26,7 @@ pub fn execute_inr(state: &mut State, inr: INR) {
         INR::IncrementMemory => {
             let address = state.get_register_pair(&RegisterPair::HL);
             let lhs = state.get_memory(address) as i8;
-            let rhs = 1 as i8;
+            let rhs = 1_i8;
 
             let (result, carried) = lhs.overflowing_add(rhs);
             let flags = get_flags(lhs, rhs, result, carried);
